renamed set_config.py -> hydra_runner.py
